Add variables for the intermediate and final outputs to go to
I do builds on an external drive because it has more space; this enables that.

On Windows, use forward slashes for path separators in these variables

It's necessary for the intermediate outputs to have a separate var from final post-processor outputs, because if keep_input_artifacts is set false, I've seen the whole directory been cleared, which includes the final output itself. So they have to be different places.